+++ crates/testkit/src/lib.rs
@@ -40,8 +40,7 @@
 //!
 //! - `datasets` (enabled by default): Enables test dataset discovery, download, validation, parsing,
 //!   and loading.
-//! - `testers` (enabled by default): Enables test actors and strategies for live testing and
-//!   development.
+//! - `testers` (enabled by default): Enables test actors, strategies, and in-memory cache backing.
 //! - `python`: Enables Python bindings from [PyO3](https://pyo3.rs).
 //! - `high-precision`: Enables [high-precision mode](https://nautilustrader.io/docs/nightly/getting_started/installation#precision-mode) to use 128-bit value types.
 //! - `extension-module`: Builds the crate as a Python extension module.
@@ -68,10 +67,14 @@
     )
 )]
 
-pub mod events;
-
+#[cfg(feature = "testers")]
+pub mod cache;
 #[cfg(feature = "datasets")]
 pub mod common;
+#[cfg(feature = "testers")]
+pub mod components;
+pub mod events;
+
 #[cfg(feature = "datasets")]
 pub mod files;
 #[cfg(feature = "datasets")]